+++ apps/edit/src/editors/PhylogenyEditor/NodesEditor/Arc.d.ts
@@ -1,3 +1,0 @@
-import { Vertex } from "./Vertex"
-
-export type Arc = Readonly<[Vertex, Vertex]>