+++ apps/www/src/tags/useTagsFilterQuery.ts
@@ -1,0 +1,11 @@
+import { ImageListParameters } from "@phylopic/api-models"
+import { useContext, useMemo } from "react"
+import TagsFilterTypeContext from "./TagsFilterTypeContext"
+const useTagsFilterQuery = () => {
+    const [tags] = useContext(TagsFilterTypeContext) ?? []
+    return useMemo<ImageListParameters>(
+        () => (tags?.size ? { filter_tags: Array.from(tags).sort().join(",") } : {}),
+        [tags],
+    )
+}
+export default useTagsFilterQuery