+++ packages/api-models/src/detection/isNode.ts
@@ -10,6 +10,8 @@ const isNodeLinks = (x: unknown, faultCollector?: ValidationFaultCollector): x i
     isNullOr(isTitledLink(isNormalizedText))((x as Node["_links"]).cladeImages, faultCollector?.sub("cladeImages")) &&
     isArray(isTitledLink(isURL))((x as Node["_links"]).external, faultCollector?.sub("external")) &&
     isNullOr(isLink(isNormalizedText))((x as Node["_links"]).images, faultCollector?.sub("images")) &&
+    // :TODO: Reinstate this after launching the update to the API
+    //isLink(isNormalizedText)((x as Node["_links"]).imageTags, faultCollector?.sub("imageTags")) &&
     isLink(isNormalizedText)((x as Node["_links"]).lineage, faultCollector?.sub("lineage")) &&
     isNullOr(isTitledLink(isNormalizedText))((x as Node["_links"]).parentNode, faultCollector?.sub("parentNode")) &&
     isNullOr(isTitledLink(isNormalizedText))((x as Node["_links"]).primaryImage, faultCollector?.sub("primaryImage"))